requirements.txt
PyPDF2==3.0.1
python-dotenv==1.0.1
pycryptodome==3.20.0